[360] refactor: split TocSection.tsx — move utilities to tocUtils.ts (fixes react-refresh lint)
