+++ packages/service/client/src/components/MarkdownView.tsx
@@ -6,7 +6,8 @@ import { useCallback, useMemo, useState } from 'react';
 import { initEmbeddedDiagramPanzoom } from '@/components/EmbeddedDiagramPanzoom';
 import { initLazyDiagrams } from '@/components/LazyDiagram';
 import { initInlineSvgPanzoom } from '@/components/InlineSvgPanzoom';
-import { TocSection, buildTocTree, findAncestorSlugs } from '@/components/TocSection';
+import { TocSection } from '@/components/TocSection';
+import { buildTocTree, findAncestorSlugs } from '@/components/tocUtils';
 import type { FileContent } from '@/lib/api';
 import { initCodeBlockCm6 } from '@/lib/codeBlockCm6';
 import { injectCopyButtons } from '@/lib/codeBlockCopy';
